requirements.txt
scapy
sqlite3
pyyaml